Moves audio stuff to separate folder
